backend/src/requirements.txt
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
oauthlib
pylint
requests
requests-oauthlib
sqlparse
urllib3
